requirements.txt
numpy==1.24.3
pypdf==3.17.4
pypdfium2==4.30.0
reportlab==4.0.7
onnxocr==2025.5
paddleocr==2.7.0.3
paddlepaddle==2.6.2
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.4.0
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.6.0.66
